created separate module for player bot (player_bot.py) added pic to readme.
