+++ README.md
@@ -48,13 +48,32 @@ To start the game:
 ```
 $ python3 game_of_greed_remastered.py
 ```
-This module includes a player bot designed to make the best choices. 
-Follow the prompts - Good luck!
+<br>**At the start of the module you will be prompted:**
+![Game_of_Greed_prompt](assets/game_of_greed_prompt.png)
+```
+To play:
+$ play
+
+To run the player bot:
+$ bot
+
+To quit:
+$ quit
+```
+**Player Bot:**
+This module includes a player bot designed to make the best choices. By electing to run the player bot, you will see the bot play through 1000 games of greed. Then you can see metrics for how the bot performed.
+
+**Play yourself:**
+By electing to play, you will be prompted to select dice to save, reroll what dice you have left, and to bank your dice. After reaching 10,000 points or after 20 rounds, the game will end.
+
+Follow the prompts - And good luck!
 
 ## Functionality/Architecture
 This game is using 'dice rolls', arguements in the form of tuples. The dice rolls are between 1 and 6 individually and there are six dice at the start of the game. An instance of a GameOfGreed is created upon game start (invoking GameOfGreed class object method play()). 
 
-User is guided through a series of prompts. They can elect to add scoring dice to their side of the table (by typing "a"), bank the points that they have set aside (by typing "b"), and/or reroll the remaining dice to go for a bigger score (by typing "r").
+User is guided through a series of prompts. They can elect to add scoring dice to their side of the table, bank the points that they have set aside, and/or reroll the remaining dice to go for a bigger score.
+
+The player bot is reading every print statement from the game of greed and is looking for two specific prints. The dice roll and the remaining dice. Given, the dice on the table, the bot makes decisions on what the best dice to keep are and whether or not to re-roll.
 
 ## Change Log
 Mon Dec 09 2019 16:34:04<br>Created Game of Greed Module. Tested for app functionality and correct results.
@@ -63,5 +82,5 @@ Tue Dec 10 2019 22:38:07<br>Developed further game functionality. Ability to rol
 
 Wed Dec 11 2019 22:19:37<br>Devloped flow testing. Ensured proper game flow.
 
-Sun Dec 29 2019 14:09:42<br>Refactored Game of Greed to incorporate a player bot easier. New module, game_of_greed_remastered.py, includes the player bot. Averaging a winning score; 510 points on average per turn.
+Sun Dec 29 2019 14:09:42<br>Refactored Game of Greed to incorporate a player bot easier. New module, game_of_greed_remastered.py, includes the player bot. Averaging a winning score; 530 points on average per turn.
